backend/requirements.txt
datasets>=2.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

